Remove the top-level fp-ts-logo.png
It is a duplicate of what is found in docs/fp-ts-logo.png.

All this moving around is not ideal, as it also breaks existing hyperlinks. But the way GitHub pages works makes it reasonable to keep things related to the documentation website under /docs so they can be referenced more easily.