modules/index.md structure suggestion
This is not mergeable as-is because it will be overwritten by the documentation generator. It is here to start a discussion on whether it would make sense to work towards something like this and whether that is actually possible.

Ideally, docs-ts would be extended to read `@tag` information or something from the source files so they can be linked grouped by tags (which in this case would be "Data Types", "Type Classes", etc.

The big benefit I would see from something like this is that it would make some useful modules more findable; I did not know about the Console module, for example, but find it useful.

The editorial content describing data types/type classes is less important and could be removed – it is covered under core-concepts.